third_party/WebKit/LayoutTests/inspector/text-autosizing-override.html

<html>
<head>
<meta name="viewport" content="width=800">
<script src="../http/tests/inspector/inspector-test.js"></script>
<style>
  html { font-size: 16px; }
  body { width: 800px; margin: 0; overflow-y: hidden; }
</style>
<script>
  if (window.internals)
    window.internals.settings.setTextAutosizingWindowSizeOverride(320, 480);

  function getAutosizing() {
    return JSON.stringify({
      textHeight: document.getElementById('measure').offsetHeight
    });
  }

  function forceAutosizing(enabled) {
    window.internals.settings.setTextAutosizingEnabled(enabled);
  }

  function test() {
    function assertAutosizingResult(expected, callback) {
      function resultCallback(jsonResult) {
        var result = JSON.parse(jsonResult.value);
        var actual = result.textHeight > 200;
        InspectorTest.addResult('Text ' + (actual ? 'was' : 'was not') + ' autosized. ' + (expected == actual ? 'PASS' : 'FAIL'));
        if (callback)
          callback();
      }
      InspectorTest.evaluateInPage("getAutosizing()", resultCallback);
    }

    InspectorTest.runTestSuite([
      function checkNotAutosizedWithoutPageAgent(next) {
        InspectorTest.evaluateInPage("forceAutosizing(false)", function() {
          assertAutosizingResult(false, next);
        });
      },

      function checkNotAutosizedWithPageAgent(next) {
        InspectorTest.evaluateInPage("forceAutosizing(false)", function() {
          PageAgent.enable();
          assertAutosizingResult(false, next);
        });
      },

      function checkAutosizedWithoutPageAgent(next) {
        InspectorTest.evaluateInPage("forceAutosizing(true)", function() {
          PageAgent.disable();
          assertAutosizingResult(true, next);
        });
      },

      function checkAutosizedWithPageAgent(next) {
        InspectorTest.evaluateInPage("forceAutosizing(true)", function() {
          PageAgent.enable();
          assertAutosizingResult(true, next);
        });
      },

      function checkAutosizedWithDeviceMetrics(next) {
        PageAgent.invoke_setDeviceMetricsOverride({width: 0, height: 0, deviceScaleFactor: 0, mobile: true, fitWindow: true}, function() {
          assertAutosizingResult(true, next);
        });
      },

      function checkNotAutosizedWithDeviceMetrics(next) {
        PageAgent.invoke_setDeviceMetricsOverride({width: 0, height: 0, deviceScaleFactor: 0, mobile: false, fitWindow: true}, function() {
          assertAutosizingResult(false, next);
        });
      },

      function checkAutosizedWithoutDeviceMetrics(next) {
        PageAgent.clearDeviceMetricsOverride(function() {
          assertAutosizingResult(true, next);
        });
      },

      function checkNotAutosizedWithoutDeviceMetrics(next) {
        InspectorTest.evaluateInPage("forceAutosizing(false)", function() {
          PageAgent.clearDeviceMetricsOverride(function() {
            assertAutosizingResult(false, next);
          });
        });
      },

      function checkAutosizedWhenDeviceMetricsOverrideSettings(next) {
        InspectorTest.evaluateInPage("forceAutosizing(false)", function() {
          PageAgent.invoke_setDeviceMetricsOverride({width: 0, height: 0, deviceScaleFactor: 0, mobile: true, fitWindow: true}, function() {
            assertAutosizingResult(true, next);
          });
        });
      },

      function cleanup(next) {
        PageAgent.clearDeviceMetricsOverride(function() {
          InspectorTest.evaluateInPage("var el = document.getElementById('measure'); el.parentNode.removeChild(el);", next);
        });
      }
    ]);
  }
</script>
</head>
<body onload="runTest()">
<div id="measure">
  This text should be autosized to 40px computed font-size (16 * 800/320).<br/>
  FAIL IF THIS IS VISIBLE<br/>
  Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
</div>
</body>
</html>
