third_party/WebKit/LayoutTests/inspector/text-autosizing-override-expected.txt


Running: checkNotAutosizedWithoutPageAgent
Text was not autosized. PASS

Running: checkNotAutosizedWithPageAgent
Text was not autosized. PASS

Running: checkAutosizedWithoutPageAgent
Text was autosized. PASS

Running: checkAutosizedWithPageAgent
Text was autosized. PASS

Running: checkAutosizedWithDeviceMetrics
Text was autosized. PASS

Running: checkNotAutosizedWithDeviceMetrics
Text was not autosized. PASS

Running: checkAutosizedWithoutDeviceMetrics
Text was autosized. PASS

Running: checkNotAutosizedWithoutDeviceMetrics
Text was not autosized. PASS

Running: checkAutosizedWhenDeviceMetricsOverrideSettings
Text was autosized. PASS

Running: cleanup
